requirements.txt
getch
python_vlc
termcolor
